Project outlining (#35)
* Added ability to generate terrain from script

TerrainGenerator script can now create a terrain object and assign the proper values to it for display in the game world

* Added basic perlin noise

the terrain is now given a 2D array of floats that give the terrain changing height using Perlin noise

* added scaling to Perlin noise

scaling affects the zooming factor on the Perlin noise.

* added ability to load materials to terrain

any materials put into the Resources folder of the project can be loaded and applied to the terrain.

* Added ability to regenerate terrain during run time

there is now a boolean option in the inspector window for the TerrainGenerator script which when enabled will regenerate the terrain with the newest settings. The terrain is different every time thanks to the perlin noise offset which acts as a random seed

* Added smooth regeneration

smooth regeneration increases the seed by a small amount each frame thus only slightly changing the current state of the terrain. This is useful for testing terrain features

* Added noise variation types

added 3 layers of noise for the base file, and experimented with a different type of Perlin noise scaling with ridge noise

* Added server handler

the server handler handles starting up the server. when you run the project it attempts to locate and run the python server.

* Added AICommunicator

AICommunicatorAlt provides similar functionality to the main one, by making a request to the server and for now generating a basic terrain if the request was successful.

* UI Fix

The text box and button now snap properly to the top left of the screen even when changing screen size or resolution

* Added JSON serialization from ai output

the output of the ai is now processed into a struct called WorldInfo that stores the parameters for world generation.

* Added Object Generation Outline

There is now the WorldGenerator class which stores all of the generator scripts. These generator scripts could be for terrain, clouds, weather effects, or even time of day.

* Minor scene update

* Updated project structure

scenes and scripts are not in their own folders

* Updated json serializer

Now using the more robust newtonsoft json serializer which can handle more than primitive data types

* Added functionality to TerrainGenerator

the terrain generator will produce a random terrain with the specified terrain material specified by gemini

* Added ability to generate terrain from script

TerrainGenerator script can now create a terrain object and assign the proper values to it for display in the game world

* Added basic perlin noise

the terrain is now given a 2D array of floats that give the terrain changing height using Perlin noise

* added scaling to Perlin noise

scaling affects the zooming factor on the Perlin noise.

* added ability to load materials to terrain

any materials put into the Resources folder of the project can be loaded and applied to the terrain.

* Added ability to regenerate terrain during run time

there is now a boolean option in the inspector window for the TerrainGenerator script which when enabled will regenerate the terrain with the newest settings. The terrain is different every time thanks to the perlin noise offset which acts as a random seed

* Added smooth regeneration

smooth regeneration increases the seed by a small amount each frame thus only slightly changing the current state of the terrain. This is useful for testing terrain features

* Added noise variation types

added 3 layers of noise for the base file, and experimented with a different type of Perlin noise scaling with ridge noise

* Added server handler

the server handler handles starting up the server. when you run the project it attempts to locate and run the python server.

* Added AICommunicator

AICommunicatorAlt provides similar functionality to the main one, by making a request to the server and for now generating a basic terrain if the request was successful.

* UI Fix

The text box and button now snap properly to the top left of the screen even when changing screen size or resolution

* Added JSON serialization from ai output

the output of the ai is now processed into a struct called WorldInfo that stores the parameters for world generation.

* Added Object Generation Outline

There is now the WorldGenerator class which stores all of the generator scripts. These generator scripts could be for terrain, clouds, weather effects, or even time of day.

* Minor scene update

* Updated project structure

scenes and scripts are not in their own folders

* Updated json serializer

Now using the more robust newtonsoft json serializer which can handle more than primitive data types

* Added functionality to TerrainGenerator

the terrain generator will produce a random terrain with the specified terrain material specified by gemini

* Removed old scripts

old scripts have been replaced with new ones.

* Updating path for linux